requirements.txt
asgiref==3.6.0
decorator==5.1.1
Django==4.1.7
django-ckeditor==6.5.1
django-etc==1.4.0
django-js-asset==2.0.0
django-likes==2.0.1
django-secretballot==2.0.0
django-tinymce==3.6.1
gunicorn==20.1.0
like==1.7.1
Pillow==9.4.0
self==2020.12.3
sqlparse==0.4.3
tzdata==2022.7
